+++ SpaceCapsule/SpaceCapsule/App/AppCoordinator.swift
@@ -19,7 +19,6 @@ final class AppCoordinator: Coordinator {
     func start() {
         // TODO: 분기처리 하기!
         // moveToAuth()
-        moveToLocation()
     }
     
     private func moveToAuth() {
@@ -34,11 +33,4 @@ final class AppCoordinator: Coordinator {
     private func moveToTabBar() {
        
     }
-    
-    // MARK: 임시 ViewController - 삭제 필요
-    private func moveToLocation() {
-        let locationController = LocationViewController()
-        rootViewController = locationController
-    }
-    
 }
