+++ SpaceCapsule/SpaceCapsule/Scene/TabBar/CapsuleAdd/CapsuleLocate/CapsuleLocateView.swift
@@ -5,13 +5,12 @@
 //  Created by 장재훈 on 2022/11/15.
 //
 
-
 import UIKit
 import SnapKit
 import MapKit
 
 final class CapsuleLocateView: UIView, BaseView {
-
+    
     var map: MKMapView = {
         let mapView = MKMapView()
         return mapView
@@ -25,35 +24,32 @@ final class CapsuleLocateView: UIView, BaseView {
     }()
     
     // MARK: - Lifecycle
-
+    
     override init(frame: CGRect) {
         super.init(frame: frame)
         configure()
         addSubViews()
         makeConstraints()
     }
-
+    
     @available(*, unavailable)
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-
+    
     // MARK: - Methods
-
+    
     func configure() {
         backgroundColor = .themeBackground
-      
     }
-
     
     func addSubViews() {
         [map, cursor].forEach {
             addSubview($0)
         }
     }
-
+    
     func makeConstraints() {
-        
         map.snp.makeConstraints {
             $0.edges.equalToSuperview()
         }
@@ -62,8 +58,5 @@ final class CapsuleLocateView: UIView, BaseView {
             $0.center.equalToSuperview()
             $0.height.width.equalTo(20)
         }
-        
     }
 }
-
-
